добавление и изменение классов AccountOperationDto, AnEntityWithAnIdOnlyDto, CreateParkingSpaceDto, CarDto, ExtendOrderDto, InfoParkingSpaceDto, OrderCreateDto, OrderInfoDto, TerminateOrderDto, UserInfoDto. Добавлены enum DepositDtoSlotBalance и OrderSlotStatus. Переделаны сущности Order и ParkingSpace. Были изменены 3 репозитория CarRepository, OrderRepository, ParkingRepository. Так же была проведена колосальная работа с сервисами CarService, OrderAndUserUtilsService, OrderService, ParkingService, UserService и польностью переделан контроллер ConsumerController и изменен AdminController

+++ RETURN/src/main/java/com/example/RETURN/repositories/ParkingRepository.java
@@ -6,8 +6,9 @@ import com.example.RETURN.models.ParkingSpace;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface ParkingRepository extends JpaRepository<ParkingSpace, Long> {
+public interface ParkingRepository extends JpaRepository<ParkingSpace, Integer> {
 
     boolean existsByParkingSlotNumber(ParkingSlotNumber number);
